move HTTP request to own module
